move actions for stories into use server file
